Move the LineSplitting module to a separate file
