+++ src/GeoMakie.jl
@@ -40,6 +40,7 @@ include("geojson.jl") # GeoJSON/GeoInterface support
 include("conversions.jl")
 include("data.jl")
 include("utils.jl")
+include("linesplitting.jl")
 include("geodesy.jl")
 
 @reexport using Colors, Makie
